Implement the ignoreId flag for class definition imports
By default ignore the class ID when provided in the .json export of the
class definition. Otherwise this might throw exceptions when different
environments have other ID's. As the classes can be generated by
scripts, the autoincrement ID isn't necessarily the same.

This change allows the ignoreId to be passed along from within the
trait. By default however it will be set to true instead of false.

Tickets: PIMCSE-23